refactor: extract file system tool callbacks in dedicated files (search/find/locate)
